move dynamic components to the external lib
