+++ src/app/components/context-menu/context-menu.component.html
@@ -37,7 +37,7 @@
             </ng-container>
 
             <ng-container *ngSwitchCase="'custom'">
-                <app-custom-component [id]="entry.component"></app-custom-component>
+                <adf-dynamic-component [id]="entry.component"></adf-dynamic-component>
             </ng-container>
         </ng-container>
     </div>
