+++ src/app/components/context-menu/context-menu.module.ts
@@ -33,6 +33,7 @@ import { ContextActionsDirective } from './context-menu.directive';
 import { ContextMenuService } from './context-menu.service';
 import { ContextMenuComponent } from './context-menu.component';
 import { ContextMenuItemDirective } from './context-menu-item.directive';
+import { ExtensionsModule } from '@alfresco/adf-extensions';
 
 @NgModule({
     imports: [
@@ -42,7 +43,8 @@ import { ContextMenuItemDirective } from './context-menu-item.directive';
         MatButtonModule,
         BrowserModule,
         CoreExtensionsModule.forChild(),
-        CoreModule.forChild()
+        CoreModule.forChild(),
+        ExtensionsModule.forChild()
     ],
     declarations: [
         ContextActionsDirective,
